+++ internal/provider/project-ds.go
@@ -1,0 +1,112 @@
+package provider
+
+import (
+	cloudv1 "buf.build/gen/go/gportal/gportal-cloud/protocolbuffers/go/gpcloud/api/cloud/v1"
+	"context"
+	"fmt"
+	"github.com/G-PORTAL/gpcloud-go/pkg/gpcloud/client"
+	"github.com/G-PORTAL/terraform-provider-gpcloud/internal/gpcloudvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+var _ datasource.DataSource = &ProjectDataSource{}
+
+func NewProjectDS() datasource.DataSource {
+	return &ProjectDataSource{}
+}
+
+// ProjectDataSource defines the data source implementation.
+type ProjectDataSource struct {
+	client *client.Client
+}
+
+// ProjectDataSourceModel describes the project data model.
+type ProjectDataSourceModel struct {
+	Name        types.String `tfsdk:"name"`
+	Description types.String `tfsdk:"description"`
+	Environment types.String `tfsdk:"environment"`
+	Id          types.String `tfsdk:"id"`
+}
+
+func (d *ProjectDataSource) Metadata(ctx context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_project"
+}
+
+func (d *ProjectDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: "The datasource project represents a project managed inside the GPCloud panel.\n\n" +
+			"Since the default authentication method does not support updating / creating projects, the datasource should\n" +
+			"be used. In case a direct access grant authentication method is available, you may want to use the `gpcloud_project` resource instead.",
+
+		Attributes: map[string]schema.Attribute{
+			"name": schema.StringAttribute{
+				MarkdownDescription: "Name of the Project",
+				Computed:            true,
+			},
+			"description": schema.StringAttribute{
+				MarkdownDescription: "Description of the Project",
+				Computed:            true,
+			},
+			"environment": schema.StringAttribute{
+				MarkdownDescription: "Project Environment",
+				Computed:            true,
+			},
+			"id": schema.StringAttribute{
+				Required:            true,
+				MarkdownDescription: "Project ID",
+				Validators: []validator.String{
+					gpcloudvalidator.UUIDStringValidator{},
+				},
+			},
+		},
+	}
+}
+
+func (d *ProjectDataSource) Configure(ctx context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	client, ok := req.ProviderData.(*client.Client)
+
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Resource Configure Type",
+			fmt.Sprintf("Expected *http.Client, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+		)
+
+		return
+	}
+
+	d.client = client
+}
+
+func (d *ProjectDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var data ProjectDataSourceModel
+	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	projectResponse, err := d.client.CloudClient().GetProject(context.Background(), &cloudv1.GetProjectRequest{
+		Id: data.Id.ValueString(),
+	})
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to get project, got error: %s", err))
+		return
+	}
+	data.write(projectResponse.Project)
+	resp.Diagnostics.AddError("Project not found", fmt.Sprintf("Project %s not found", data.Id.ValueString()))
+}
+
+func (data *ProjectDataSourceModel) write(project *cloudv1.Project) {
+	data.Id = types.StringValue(project.Id)
+	data.Name = types.StringValue(project.Name)
+	data.Environment = types.StringValue(project.Environment.String())
+	if project.Description != "" {
+		data.Description = types.StringValue(project.Description)
+	}
+}